Ensure plugin load methods are called only when path/match/query changes and ignore the rest of query-params
