+++ src/common/utils/index.ts
@@ -8,3 +8,4 @@ export {
   warmupMatchRouteCache,
 } from './match-route';
 export { findRouterContext, createRouterContext } from './router-context';
+export { isSameRoute } from './is-same-route';
